requirements.txt
numpy==1.26.4
scipy==1.12.0
pyserial==3.5
pyyaml==6.0.1
psutil==5.9.8
pytest==8.1.0
